Use `virtualInterface`.
